strideguard/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
opencv-python
pytesseract
Pillow
openai
